Update bible_reading_plan.py
I have updated the file and commented out the books of the Bible that I have read.  Now anyone can access this file and they can take the code that I have provided.  All they need to do is uncomment the books of the Bible that they have read so that they can keep track.